pyproject.toml
[project]
name = "library-mannagment-system-phitron"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django>=5.2.4",
    "django-rest-framework>=0.1.0",
    "djangorestframework>=3.16.0",
    "djangorestframework-simplejwt>=5.5.1",
    "djoser>=2.3.3",
    "drf-yasg>=1.21.10",
]
